refactor: extract summary and memory draft requests
