+++ src/core/services/generationService.js
@@ -35,6 +35,8 @@ import {
 import { runChatPostPromptPipeline } from '@/core/llm/usecases/chatPostPromptPipeline.js';
 import { executePreparedChatPrompt } from '@/core/llm/usecases/chatPreparedPromptExecution.js';
 import { executeChatContextCalculation } from '@/core/llm/usecases/chatContextCalculation.js';
+import { executeSummaryRequest } from '@/core/llm/usecases/summaryRequest.js';
+import { executeMemoryDraftRequest } from '@/core/llm/usecases/memoryDraftRequest.js';
 
 let lastPrompt = null;
 
@@ -406,45 +408,17 @@ export async function calculateContext({ char, history, authorsNote, summary }) 
 }
 
 export async function generateSummary({ history, prompt, controller, apiConfigOverride = null }) {
-    const effectiveConfig = {
-        ...getEffectiveApiConfig(),
-        ...(apiConfigOverride || {})
-    };
-    const { providerId, apiKey, apiUrl, model, temp } = effectiveConfig;
-
-    if (!apiUrl || !model) {
-        throw new Error("API Not Configured");
-    }
-
-    const defaultPrompt = "Summarize the following roleplay conversation concisely, focusing on the current situation and key events:\n\n{{history}}";
-    const template = prompt || defaultPrompt;
-
-    let finalPrompt = template.replace('{{history}}', history);
-    if (!template.includes('{{history}}')) {
-        finalPrompt = `${template}\n\n${history}`;
-    }
-
-    let result = "";
-
-    const { requestBody } = buildSummaryRequestPayload({
-        providerId,
-        model,
-        prompt: finalPrompt,
-        temperature: temp
-    });
-
-    await executeRequest({
-        providerId,
-        apiUrl,
-        apiKey,
-        requestBody,
+    return executeSummaryRequest({
+        history,
+        prompt,
         controller,
-        callbacks: {
-            onComplete: (text) => { result = text; }
+        apiConfigOverride,
+        deps: {
+            getEffectiveApiConfig,
+            buildSummaryRequestPayload,
+            executeRequest
         }
     });
-
-    return result;
 }
 
 function normalizeMessageIdList(entry) {
@@ -732,77 +706,18 @@ async function buildMemoryInjection({ char, history, summary, safeContext, cutof
 }
 
 export async function generateMemoryDraft({ history, prompt, controller, apiConfigOverride = null }) {
-    const effectiveConfig = {
-        ...getEffectiveApiConfig(),
-        ...(apiConfigOverride || {})
-    };
-    const { providerId, apiKey, apiUrl, model, temp } = effectiveConfig;
-
-    // Memory drafts need enough output budget for long summaries even when the
-    // provider has a small default completion limit.
-    const explicitOverrideMaxTokens = Number(apiConfigOverride?.maxTokens);
-    const hasExplicitOverride = Number.isFinite(explicitOverrideMaxTokens) && explicitOverrideMaxTokens > 0;
-    const configuredMaxTokens = Number(effectiveConfig.maxTokens);
-    const memoryDraftMaxTokens = hasExplicitOverride
-        ? Math.max(200, Math.round(explicitOverrideMaxTokens))
-        : (Number.isFinite(configuredMaxTokens) && configuredMaxTokens > 0
-            ? Math.max(1200, Math.round(configuredMaxTokens))
-            : 2000);
-
-    if (!apiUrl || !model) {
-        throw new Error("API Not Configured");
-    }
-
-    const defaultPrompt = [
-        'Create exactly one concise long-term memory entry from the following roleplay segment.',
-        'Preserve the original language of the source segment. Do not translate it.',
-        'Use only facts that are explicitly supported by the segment.',
-        'Do not infer completed outcomes, registrations, approvals, or decisions unless the text clearly states them.',
-        'Focus on durable facts, developments, or relationship changes that should persist beyond immediate context.',
-        'Do not copy the dialogue verbatim.',
-        'Return only the memory entry text with no preface, label, or explanation.',
-        '',
-        '{{history}}'
-    ].join('\n');
-    const template = prompt || defaultPrompt;
-
-    let finalPrompt = template.replace('{{history}}', history);
-    if (!template.includes('{{history}}')) {
-        finalPrompt = `${template}\n\n${history}`;
-    }
-
-    let result = "";
-    
-    let requestError = null;
-    const { previewBody, requestBody } = buildMemoryDraftRequestPayload({
-        providerId,
-        model,
-        prompt: finalPrompt,
-        temperature: temp,
-        maxTokens: memoryDraftMaxTokens
-    });
-
-    lastPrompt = JSON.parse(JSON.stringify(previewBody));
-    
-    await executeRequest({
-        providerId,
-        apiUrl,
-        apiKey,
-        requestBody,
-        stream: false,
+    return executeMemoryDraftRequest({
+        history,
+        prompt,
         controller,
-        requestType: 'memory_draft',
-        callbacks: {
-            onUpdate: (chunk, reasoningChunk, effectiveText) => {
-                if (effectiveText) result = effectiveText;
-                else if (chunk) result += chunk;
-            },
-            onComplete: (text) => { if (text) result = text; },
-            onError: (err) => { requestError = err; }
+        apiConfigOverride,
+        deps: {
+            getEffectiveApiConfig,
+            buildMemoryDraftRequestPayload,
+            executeRequest,
+            setLastPrompt: (promptBody) => {
+                lastPrompt = promptBody;
+            }
         }
     });
-    
-    if (requestError) throw requestError;
-
-    return result;
 }
